Improved interthread synchronisation with RMI worker threads in PacBioKanga, increased number of parsed fields to represent actual number of cultivars when generating SNP marker SQLite database in BioKanga. PacBioKanga release is now 1.2.4, and BioKanga is now 3.9.5

+++ biokanga/Markers.cpp
@@ -1067,7 +1067,7 @@ CMarkers::IdentSpeciesSpec(int AltMaxCnt,	// max count allowed for base being pr
 						int MinSpeciesTotCntThres)		// individual species must have at least this number of total bases at SNP loci - 0 if no threshold
 
 {
-UINT32 AlignIdx;
+INT64 AlignIdx;
 tsAlignLoci *pAlign;
 tsAlignLoci *pAlignSpecies;
 tsAlignLoci *pAlignSpeciesA;
